refactor: extract PartialCursor record
